backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
mysql-connector-python==9.1.0